services/humanizer-svc/pyproject.toml
[project]
name = "studojo-humanizer"
version = "0.1.0"
description = "Structure-preserving document humanization service"
readme = "README.rst"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-docx>=1.2.0",
    "requests>=2.32.0",
    "openai>=2.8.0",
    "pydantic>=2.12.4",
    "python-multipart>=0.0.12",
]

